+++ src/library/DIPS.Mobile.UI/Resources/Styles/Button/ButtonStyle.cs
@@ -8,5 +8,11 @@ public enum ButtonStyle
     SecondaryLarge = 3,
     SecondarySmall = 4,
     GhostLarge = 5,
-    GhostSmall = 6
+    GhostSmall = 6,
+    PrimaryIconButtonSmall = 7,
+    PrimaryIconButtonLarge = 8,
+    SecondaryIconButtonSmall = 9,
+    SecondaryIconButtonLarge = 10,
+    GhostIconButtonSmall = 11,
+    GhostIconButtonLarge = 12,
 }