+++ CHANGELOG.md
@@ -1,6 +1,17 @@
-## [17.19.5]
+## [17.20.3]
 - Added ContextMenuOptions on ListItem
 - [Android] Fixed an issue where setting both ContextMenu and Touch API, would disable ContextMenu
+
+## [17.20.2]
+- [Button] Renamed Rounded styles to IconButton
+
+## [17.20.1]
+- [Button] Changed the disabled text color of buttons.
+
+## [17.20.0]
+- [Button][Styles] Added PrimaryRoundedSmall/Large
+- [Button][Styles] Added SecondaryRoundedSmall/Large
+- [Button][Styles] Added GhostRoundedSmall/Large
 
 ## [17.19.4]
 - [iOS] Fixed another issue where Context menu would crash when in a CollectionView